Disambiguate uses of sqrt by casting to double.
Should fix a a Visual Studio compile error (felirx).

+++ crawl-ref/source/dgn-shoals.cc
@@ -886,7 +886,7 @@ static int _shoals_tide_at(coord_def pos, int base_tide)
         return base_tide;
 
     const int distance =
-        static_cast<int>(sqrt((pos - tide_caller->pos()).abs()));
+        static_cast<int>(sqrt((double)(pos - tide_caller->pos()).abs()));
     if (distance > TIDE_CALL_RADIUS)
         return base_tide;
